Renamed model package in Shared module
It turns out we can name packages the same for the Server and app
modules, but the Shared packages need to have unique names
