--- TicketToRide/Shared/src/main/java/model/User.java
@@ -1,16 +1,0 @@
-package main.java.model;
-
-import java.util.ArrayList;
-
-public class User {
-    private String username;
-    private ArrayList<GameID> gameIDlist;
-
-    public User(String username) {
-        this.username = username;
-    }
-
-    public String getUsername() {
-        return username;
-    }
-}
