+++ TicketToRide/Shared/src/main/java/modelclasses/GameID.java
@@ -1,4 +1,4 @@
-package main.java.model;
+package modelclasses;
 
 public class GameID {
 }
