+++ TicketToRide/Shared/src/main/java/modelclasses/GameInfo.java
@@ -1,4 +1,4 @@
-package main.java.model;
+package modelclasses;
 
 import java.util.ArrayList;
