+++ TicketToRide/Shared/src/main/java/modelclasses/PlayerColor.java
@@ -1,4 +1,4 @@
-package main.java.model;
+package modelclasses;
 
 public class PlayerColor {
 }
