+++ TicketToRide/Shared/src/main/java/modelclasses/User.java
@@ -1,0 +1,16 @@
+package modelclasses;
+
+import java.util.ArrayList;
+
+public class User {
+    private String username;
+    private ArrayList<GameID> gameIDlist;
+
+    public User(String username) {
+        this.username = username;
+    }
+
+    public String getUsername() {
+        return username;
+    }
+}
